+++ src/main/java/com/qa/InspectorFacade/GogoGadgetMongoClientServiceApplication.java
@@ -1,4 +1,4 @@
-package com.qa.gogoGadgetMongoClientService;
+package com.qa.InspectorFacade;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
