+++ src/main/java/com/qa/InspectorFacade/persistence/domain/ReflectionForm.java
@@ -1,10 +1,15 @@
-package com.qa.gogoGadgetMongoClientService.persistence.domain;
+package com.qa.InspectorFacade.persistence.domain;
 
-public class SentReflectionForm {
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.Id;
 
-private Long formId;
-	
-	private Long traineeId;
+@Entity
+public class ReflectionForm {
+
+	@Id
+	@GeneratedValue
+	private Long formId;
 	
 	private int weekNumber;
 	
@@ -18,18 +23,21 @@ private Long formId;
 	
 	private String howToStopDoingBad;
 	
-	public SentReflectionForm() {
+	private Long traineeId;
+	
+	public ReflectionForm() {
 		
 	}
 	
-	public SentReflectionForm(ReflectionForm reflectionForm) {
-		this.weekNumber = reflectionForm.getWeekNumber();
-		this.howsYourWeek = reflectionForm.getHowsYourWeek();
-		this.whatWentWell = reflectionForm.getWhatWentWell();
-		this.howToKeepDoingWell = reflectionForm.getHowToKeepDoingWell();
-		this.whatWentBad = reflectionForm.getWhatWentBad();
-		this.howToStopDoingBad = reflectionForm.getHowToStopDoingBad();
-		this.traineeId = reflectionForm.getTraineeId();
+	public ReflectionForm(int weekNumber, int howsYourWeek, String whatWentWell, String howToKeepDoingWell,
+			String whatWentBad, String howToStopDoingBad, Long traineeId) {
+		this.setWeekNumber(weekNumber);
+		this.setHowsYourWeek(howsYourWeek);
+		this.setWhatWentWell(whatWentWell);
+		this.setHowToKeepDoingWell(howToKeepDoingWell);
+		this.setWhatWentBad(whatWentBad);
+		this.setHowToStopDoingBad(howToStopDoingBad);
+		this.traineeId = traineeId;
 	}
 
 	public int getHowsYourWeek() {
