+++ src/main/java/com/qa/InspectorFacade/persistence/domain/SentCohort.java
@@ -1,4 +1,4 @@
-package com.qa.gogoGadgetMongoClientService.persistence.domain;
+package com.qa.InspectorFacade.persistence.domain;
 
 public class SentCohort {
