+++ src/main/java/com/qa/InspectorFacade/persistence/domain/SentReflectionForm.java
@@ -1,15 +1,10 @@
-package com.qa.gogoGadgetMongoClientService.persistence.domain;
+package com.qa.InspectorFacade.persistence.domain;
 
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.Id;
+public class SentReflectionForm {
 
-@Entity
-public class ReflectionForm {
-
-	@Id
-	@GeneratedValue
-	private Long formId;
+private Long formId;
+	
+	private Long traineeId;
 	
 	private int weekNumber;
 	
@@ -23,21 +18,18 @@ public class ReflectionForm {
 	
 	private String howToStopDoingBad;
 	
-	private Long traineeId;
-	
-	public ReflectionForm() {
+	public SentReflectionForm() {
 		
 	}
 	
-	public ReflectionForm(int weekNumber, int howsYourWeek, String whatWentWell, String howToKeepDoingWell,
-			String whatWentBad, String howToStopDoingBad, Long traineeId) {
-		this.setWeekNumber(weekNumber);
-		this.setHowsYourWeek(howsYourWeek);
-		this.setWhatWentWell(whatWentWell);
-		this.setHowToKeepDoingWell(howToKeepDoingWell);
-		this.setWhatWentBad(whatWentBad);
-		this.setHowToStopDoingBad(howToStopDoingBad);
-		this.traineeId = traineeId;
+	public SentReflectionForm(ReflectionForm reflectionForm) {
+		this.weekNumber = reflectionForm.getWeekNumber();
+		this.howsYourWeek = reflectionForm.getHowsYourWeek();
+		this.whatWentWell = reflectionForm.getWhatWentWell();
+		this.howToKeepDoingWell = reflectionForm.getHowToKeepDoingWell();
+		this.whatWentBad = reflectionForm.getWhatWentBad();
+		this.howToStopDoingBad = reflectionForm.getHowToStopDoingBad();
+		this.traineeId = reflectionForm.getTraineeId();
 	}
 
 	public int getHowsYourWeek() {
