+++ src/main/java/com/qa/InspectorFacade/persistence/domain/Trainee.java
@@ -1,4 +1,4 @@
-package com.qa.gogoGadgetMongoClientService.persistence.domain;
+package com.qa.InspectorFacade.persistence.domain;
 
 
 import javax.persistence.Entity;
