+++ src/main/java/com/qa/InspectorFacade/persistence/repository/MongoFormRepo.java
@@ -1,9 +1,9 @@
-package com.qa.gogoGadgetMongoClientService.persistence.repository;
+package com.qa.InspectorFacade.persistence.repository;
 
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
 
-import com.qa.gogoGadgetMongoClientService.persistence.domain.SentReflectionForm;
+import com.qa.InspectorFacade.persistence.domain.SentReflectionForm;
 
 @Repository
 public interface MongoFormRepo extends MongoRepository<SentReflectionForm, Long> {
