+++ src/main/java/com/qa/InspectorFacade/service/TraineeService.java
@@ -1,8 +1,8 @@
-package com.qa.gogoGadgetMongoClientService.service;
+package com.qa.InspectorFacade.service;
 
 import java.util.ArrayList;
 
-import com.qa.gogoGadgetMongoClientService.persistence.domain.SentTrainee;
+import com.qa.InspectorFacade.persistence.domain.SentTrainee;
 
 public interface TraineeService {
